✨feat: #14 Header 브레이킹포인트 이전 반응형 작업 및 컴포넌트 분리 작업
