+++ src/app/css/basic_mediaQuery.css
@@ -1,75 +1,79 @@
-body::-webkit-scrollbar,
-article::-webkit-scrollbar {
-  width:5px;
-}
-body::-webkit-scrollbar-thumb,
-article::-webkit-scrollbar-thumb {
-  background:#555;
-  border-radius:3px;
-}
-article::-webkit-scrollbar-corner {
-  background:transparent;
-}
+@media screen and (max-width: 1366px) {
+  header {
+    background:blue;
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 9999;
+    width: 100vw;
+    height: 100px;
+    font-weight: 300;
+    font-size: 20px;
+    color: #bbb;
+  }
+  header::after {
+    display: block;
+    content: '';
+    z-index: -1;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: #000;
+    opacity: 100%;
+  }
+  header.mainHd::after {
+    opacity: 60%;
+  }
+  
+  header .gnb_logo {
+    width: 200px;
+  }
+  header .gnb_logo a img {
+    width: 100%;
+    object-fit: cover;
+  }
 
-.subGnb::-webkit-scrollbar {
-  display: none;
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-}
-
-/* 컴포넌트 분할 */
-button {
-  border: 1px solid #777;
-  background: inherit;
-  font: 500 14px/1 'Hyundai Sans Head Office';
-  cursor: pointer;
-}
-.mainBtn {
-  background: transparent;
-  border: 1px solid #fff;
-  color: #fff;
-  width: 240px;
-  height: 60px;
-  font: 500 20px/1 'Hyundai Sans Head Office';
-  transition: all 0.2s;
-}
-.mainBtn.kr {
-  font:500 20px/1 'Pretendard'
-}
-a.mainBtn {
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  header .gnb {
+    position: absolute;
+    transform: translate(-50%,-50%);
+    top: 50%;
+    left:50%;
+    padding: 0 5rem;
+    display: none;
+    justify-content: space-between;
+    width: 100vw;
+  }
+  header .gnb > li > ul {
+    display: flex;
+    align-items: center;
+    height: 100%;
+    gap: 80px;
+  }
+  header .gnb > li:nth-child(2) {
+    align-content: end;
+    font-size: 16px;
+    justify-self: end;
+  }
+  header .gnb > li:nth-child(2) ul {
+    gap: 10px;
+  }
+  header .gnb > li:nth-child(2) > ul > li:first-child::after {
+    content: '|';
+    font-size: 10px;
+    padding-left: 10px;
+  }
+  header .gnb > li:nth-child(2) > ul > li:last-child {
+    padding-left: 80px;
+  }
+  header .smBtn {
+    justify-self: center;
+    min-width:30px;
+  }
+  
   
 }
-.mainBtn:hover {
-  background:#fff;
-  border-color:transparent;
-  color:#000;
-}
-.btnBasic {
-  border: 1px solid #000;
-  color:#fff;
-  background:#000;
-  padding: 0.7% 3%;
-  transition: 0.2s all;
-  display:flex;
-  align-items: center;
-  justify-content: center;
-}
-.btnBasic:hover{
-  background:#555;
-}
-
-/* swiper 색상 변경 */
-.swiper-button-prev,
-.swiper-button-next {
-  color: #aaa !important;
-}
-.swiper-pagination-bullet {
-  background: #fff !important;
-}
-/* 다른페이지가 붙을경우 헤딩 변경 */
 
 body {
   position: relative;
@@ -84,173 +88,7 @@ body.otherPage header::after {
   opacity: 100%;
 }
 /* Header */
-header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 9999;
-  width: 100vw;
-  height: 100px;
-  font-weight: 300;
-  font-size: 20px;
-  color: #bbb;
-}
-header::after {
-  display: block;
-  content: '';
-  z-index: -1;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: #000;
-  opacity: 100%;
-}
-header.mainHd::after {
-  opacity: 60%;
-}
 
-header .gnb_logo {
-  width: 200px;
-}
-header .gnb_logo a img {
-  width: 100%;
-  object-fit: cover;
-}
-header .gnb {
-  position: absolute;
-  transform: translate(-50%,-50%);
-  top: 50%;
-  left:50%;
-  padding: 0 5rem;
-  display: flex;
-  justify-content: space-between;
-  width: 100vw;
-}
-header .gnb > li > ul {
-  display: flex;
-  align-items: center;
-  height: 100%;
-  gap: 80px;
-}
-header .gnb > li:nth-child(2) {
-  align-content: end;
-  font-size: 16px;
-  justify-self: end;
-}
-header .gnb > li:nth-child(2) ul {
-  gap: 10px;
-}
-header .gnb > li:nth-child(2) > ul > li:first-child::after {
-  content: '|';
-  font-size: 10px;
-  padding-left: 10px;
-}
-header .gnb > li:nth-child(2) > ul > li:last-child {
-  padding-left: 80px;
-}
-header .smBtn {
-  justify-self: center;
-  min-width:30px;
-}
-
-header .modal {
-  width: 100vw;
-  height: 100vh;
-  background: #000;
-  position: fixed;
-  top: -500%;
-  opacity: 0%;
-  transition: 0.3s ease-in-out;
-  z-index: -1;
-}
-header .modal.on {
-  opacity: 95%;
-  top: 0;
-  z-index: 9999;
-}
-header .modal .closeBtn {
-  position: absolute;
-  right: 80px;
-  top: 40px;
-  width: 30px;
-  height: 30px;
-  cursor: pointer;
-}
-
-header .modal section {
-  width: 80vw;
-  height: 90vh;
-  margin: 40px auto;
-  position:relative;
-}
-header .modal section .siteItemWrap {
-  display: flex;
-  flex-direction: column;
-  row-gap: 4rem;
-}
-header .modal section h2 {
-  font: 50px/1 'Hyundai Sans Head Office';
-  color: #fff;
-  text-align: center;
-  margin-bottom: 100px;
-}
-header .modal section h3 {
-  font-size: 25px;
-  color: #fff;
-  position: relative;
-}
-header .modal section h3:after {
-  position: absolute;
-  right: 0px;
-  width: 100%;
-  height: 1px;
-  background: #666;
-  display: block;
-  content: '';
-}
-header .modal section .siteItemWrap article div {
-  padding-left: 20px;
-}
-header .modal section .siteItemWrap article div a {
-  align-self: center;
-  transition: all 0.1s;
-}
-header .modal section .siteItemWrap article div a:hover {
-  color: #fff;
-}
-header .modal section .siteItemWrap article:nth-child(1) div {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  grid-auto-rows: 50px;
-}
-header .modal section .siteItemWrap article:nth-child(2)>div {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  grid-auto-rows: 50px;
-}
-header .modal section .siteItemWrap article:nth-child(3) div {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  grid-auto-rows: 50px;
-}
-
-
-
-header .modal .closeBtn .cl-line {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 100%;
-  height: 2px;
-  background: #fff;
-  transform-origin: center;
-  transform: rotate(45deg);
-}
-header .modal .closeBtn .cl-line:last-child {
-  transform: rotate(-45deg);
-}
 
 .main_pagenation {
   position: fixed;
