+++ src/components/mainPageComponents/Event1.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { DOMElement, ReactElement, useEffect, useRef } from "react"
+import { useEffect, useRef } from "react"
 
 export default function Event1 () {
